+++ client_with_custom_token_lifespans_test.go
@@ -19,15 +19,29 @@
  *
  */
 
-package openid
+package fosite
 
 import (
-	"context"
+	"testing"
 	"time"
 
-	"github.com/ory/fosite"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-type OpenIDConnectTokenStrategy interface {
-	GenerateIDToken(ctx context.Context, lifespan time.Duration, requester fosite.Requester) (token string, err error)
+func TestDefaultClientWithCustomTokenLifespans(t *testing.T) {
+	clc := &DefaultClientWithCustomTokenLifespans{
+		DefaultClient: &DefaultClient{},
+	}
+
+	assert.Equal(t, clc.GetTokenLifespans(), (*ClientLifespanConfig)(nil))
+
+	require.Equal(t, time.Minute*42, GetEffectiveLifespan(clc, GrantTypeImplicit, IDToken, time.Minute*42))
+
+	customLifespan := 36 * time.Hour
+	clc.SetTokenLifespans(&ClientLifespanConfig{ImplicitGrantIDTokenLifespan: &customLifespan})
+	assert.NotEqual(t, clc.GetTokenLifespans(), nil)
+
+	require.Equal(t, customLifespan, GetEffectiveLifespan(clc, GrantTypeImplicit, IDToken, time.Minute*42))
+	var _ ClientWithCustomTokenLifespans = clc
 }