+++ handler/openid/helper_test.go
@@ -25,6 +25,7 @@ import (
 	"context"
 	"net/url"
 	"testing"
+	"time"
 
 	"github.com/ory/fosite/internal/gen"
 
@@ -74,19 +75,19 @@ func TestGenerateIDToken(t *testing.T) {
 			setup: func() {
 				ar.Form.Set("nonce", "11111111111111111111111111111111111")
 				ar.SetSession(sess)
-				chgen.EXPECT().GenerateIDToken(nil, ar).Return("", fooErr)
+				chgen.EXPECT().GenerateIDToken(nil, time.Duration(0), ar).Return("", fooErr)
 			},
 			expectErr: fooErr,
 		},
 		{
 			description: "should pass",
 			setup: func() {
-				chgen.EXPECT().GenerateIDToken(nil, ar).AnyTimes().Return("asdf", nil)
+				chgen.EXPECT().GenerateIDToken(nil, time.Duration(0), ar).AnyTimes().Return("asdf", nil)
 			},
 		},
 	} {
 		c.setup()
-		token, err := h.generateIDToken(nil, ar)
+		token, err := h.generateIDToken(nil, time.Duration(0), ar)
 		assert.True(t, err == c.expectErr, "(%d) %s\n%s\n%s", k, c.description, err, c.expectErr)
 		if err == nil {
 			assert.NotEmpty(t, token, "(%d) %s", k, c.description)
@@ -109,7 +110,7 @@ func TestIssueExplicitToken(t *testing.T) {
 
 	resp.EXPECT().SetExtra("id_token", gomock.Any())
 	h := &IDTokenHandleHelper{IDTokenStrategy: strat}
-	err := h.IssueExplicitIDToken(nil, ar, resp)
+	err := h.IssueExplicitIDToken(nil, time.Duration(0), ar, resp)
 	assert.NoError(t, err)
 }
 
@@ -126,7 +127,7 @@ func TestIssueImplicitToken(t *testing.T) {
 
 	resp.EXPECT().AddParameter("id_token", gomock.Any())
 	h := &IDTokenHandleHelper{IDTokenStrategy: strat}
-	err := h.IssueImplicitIDToken(nil, ar, resp)
+	err := h.IssueImplicitIDToken(nil, time.Duration(0), ar, resp)
 	assert.NoError(t, err)
 }