+++ integration/client_credentials_grant_test.go
@@ -24,21 +24,23 @@ package integration_test
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/tidwall/gjson"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/tidwall/gjson"
 	goauth "golang.org/x/oauth2"
 
 	"github.com/ory/fosite"
 	"github.com/ory/fosite/compose"
 	"github.com/ory/fosite/handler/oauth2"
+	"github.com/ory/fosite/internal"
 )
 
 func TestClientCredentialsFlow(t *testing.T) {
@@ -69,6 +71,8 @@ func runClientCredentialsGrantTest(t *testing.T, strategy oauth2.AccessTokenStra
 	defer ts.Close()
 
 	oauthClient := newOAuth2AppClient(ts)
+	fositeStore.Clients["my-client"].(*fosite.DefaultClient).RedirectURIs[0] = ts.URL + "/callback"
+	fositeStore.Clients["custom-lifespan-client"].(*fosite.DefaultClientWithCustomTokenLifespans).RedirectURIs[0] = ts.URL + "/callback"
 	for k, c := range []struct {
 		description string
 		setup       func()
@@ -113,6 +117,34 @@ func runClientCredentialsGrantTest(t *testing.T, strategy oauth2.AccessTokenStra
 				introspect(t, ts, token.AccessToken, &j, oauthClient.ClientID, oauthClient.ClientSecret)
 				assert.Equal(t, oauthClient.ClientID, gjson.GetBytes(j, "client_id").String())
 				assert.Equal(t, "fosite", gjson.GetBytes(j, "scope").String())
+				atReq, ok := fositeStore.AccessTokens[strings.Split(token.AccessToken, ".")[1]]
+				require.True(t, ok)
+				atExp := atReq.GetSession().GetExpiresAt(fosite.AccessToken)
+				internal.RequireEqualTime(t, time.Now().UTC().Add(time.Hour), atExp, time.Minute)
+				atExpIn := time.Duration(token.Extra("expires_in").(float64)) * time.Second
+				internal.RequireEqualDuration(t, time.Hour, atExpIn, time.Minute)
+			},
+		},
+		{
+			description: "should pass with custom client token lifespans",
+			setup: func() {
+				oauthClient.ClientID = "custom-lifespan-client"
+			},
+			check: func(t *testing.T, token *goauth.Token) {
+				var j json.RawMessage
+				introspect(t, ts, token.AccessToken, &j, oauthClient.ClientID, oauthClient.ClientSecret)
+				introspect(t, ts, token.AccessToken, &j, oauthClient.ClientID, oauthClient.ClientSecret)
+				assert.Equal(t, oauthClient.ClientID, gjson.GetBytes(j, "client_id").String())
+				assert.Equal(t, "fosite", gjson.GetBytes(j, "scope").String())
+
+				atReq, ok := fositeStore.AccessTokens[strings.Split(token.AccessToken, ".")[1]]
+				require.True(t, ok)
+				atExp := atReq.GetSession().GetExpiresAt(fosite.AccessToken)
+				internal.RequireEqualTime(t, time.Now().UTC().Add(*internal.TestLifespans.ClientCredentialsGrantAccessTokenLifespan), atExp, time.Minute)
+				atExpIn := time.Duration(token.Extra("expires_in").(float64)) * time.Second
+				internal.RequireEqualDuration(t, *internal.TestLifespans.ClientCredentialsGrantAccessTokenLifespan, atExpIn, time.Minute)
+				rtExp := atReq.GetSession().GetExpiresAt(fosite.RefreshToken)
+				internal.RequireEqualTime(t, time.Time{}, rtExp, time.Minute)
 			},
 		},
 	} {